fix: initualize TEAM_SIZE as usize to avoid casting

+++ src/constants.rs
@@ -2,7 +2,7 @@ use ggez::graphics::Color;
 
 pub const ENABLE_VSYNC: bool = true;
 
-pub const TEAM_SIZE: u8 = 1;
+pub const TEAM_SIZE: usize = 1;
 
 pub const VIRTUAL_WIDTH: f32 = 1980.0;
 pub const VIRTUAL_HEIGHT: f32 = 1080.0;
